fix import bug and lint
